+++ packages/esense_flutter/lib/esense.dart
@@ -91,11 +91,13 @@ class ESenseManager {
   /// Default sampling rate is 10 Hz.
   /// Returns `true` if the request was successfully made, `false` otherwise.
   Future<bool> setSamplingRate(int rate) async {
-    assert(_samplingRate > 0 && _samplingRate <= 100,
+    assert(rate > 0 && rate <= 100,
         'Must provide a sampling rate between 1 and 100 Hz.');
     _samplingRate = rate;
-    return await _eSenseManagerMethodChannel
-            .invokeMethod('setSamplingRate', <String, dynamic>{'rate': rate}) ??
+    // note that for some strange reason, iOS does not accept an int as argument
+    // hence, [rate] is converted to a string
+    return await _eSenseManagerMethodChannel.invokeMethod(
+            'setSamplingRate', <String, dynamic>{'rate': '$rate'}) ??
         false;
   }