+++ packages/health/lib/health.dart
@@ -3,9 +3,9 @@ library health;
 import 'dart:async';
 import 'dart:io' show Platform;
 
-import 'package:intl/intl.dart';
-import 'package:device_info/device_info.dart';
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/services.dart';
+import 'package:intl/intl.dart';
 
 part 'src/data_types.dart';
 part 'src/functions.dart';