+++ packages/esense_flutter/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.4.3
+* fix of error in setSamplingRate method on iOS
+
 ## 0.4.0
 * upgrade to null-safety