+++ packages/health/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.1.1+1
+* [PR #394](https://github.com/cph-cachet/flutter-plugins/pull/394) - added functions to request authorization
+
+## 3.1.0
+* [PR #372](https://github.com/cph-cachet/flutter-plugins/pull/372) - added sleep data to Android + fix of permissions and initialization
+* [PR #388](https://github.com/cph-cachet/flutter-plugins/pull/388) - testability of HealthDataPoint
+* update to using the `device_info_plus` plugin 
+
 ## 3.0.6
 * [PR#281](https://github.com/cph-cachet/flutter-plugins/pull/281) Added two new fields to the `HealthDataPoint` - `SourceId` and `SourceName` and populate when data is read. This allows datapoints to be disambigous and in some cases allows us to get more accurate data. For example the number of steps can be reported from Apple Health and Watch and without source data they are aggregated into just "steps" producing an innacurate result.